replace various jappa/kraineff textures and some other tweaks
The one set of textures I have not changed are the outdated ore textures — there seems to be a *lot* of CTM I don't want to mess with for now.

In addition, I'm not entirely sure how many of these were left in intentionally, so you may want to quickly run through the changes to make sure I haven't flagged anything intentional.

Finally, I also ran oxipng on all the directories I changed since you seem to be using lossless compression for everything.